legal_discovery_backend/requirements.txt
# Core FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0  # For PostgreSQL support

# LangGraph & AI - Use existing installed versions
# These should already be installed in your environment
# langgraph
# langchain
# langchain-core
# langchain-openai
# langchain-anthropic

# Vector Database
weaviate-client>=4.0.0,<5.0.0

# Document Processing
python-docx==1.1.0
PyPDF2==3.0.1
python-multipart==0.0.6

# Export functionality
reportlab==4.0.6

# Web and HTTP
websockets==12.0
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Optional: For enhanced document processing
# pytesseract  # OCR for images
# Pillow  # Image processing